move MainViewModel
