+++ app/src/main/java/com/example/compnote/presentation/screens/main/MainScreen.kt
@@ -20,7 +20,6 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
-import com.example.compnote.presentation.MainViewModel
 import com.example.compnote.domain.models.Note
 import com.example.compnote.presentation.navigation.NavRoute
